feast teardown deletes tables and indexes instead of dropping the schema
Previous functionality for `feast teardown` in feast-postgres was to
DROP the schema that was used.
This could cause feast-postgres to delete artifacts that were not created by
it, and users had to be aware of this.
feast-postgres will now only delete tables and indexes that belong to
the online store.